server/requirements-frozen.txt
bpython==0.12
cached-property==1.3.0
coloredlogs==5.0
Django==1.5.7
django-braces==0.2.1
django-debug-toolbar==1.2.1
django-filter==0.6
django-memcached==0.1.2
django-model-utils==1.1.0
django-registration==0.9b1
django-reversion==1.7
djangorestframework==3.1.3
elasticsearch==2.1.0
elasticsearch-dsl==0.0.9
et-xmlfile==1.0.1
fuzzywuzzy==0.8.0
gunicorn==0.17.0
humanfriendly==1.42
iso8601==0.1.11
jdcal==1.2
jsonfield==1.0.3
logutils==0.3.3
Markdown==2.3.1
MySQL-python==1.2.5
openpyxl==2.3.2
pip-accel==0.37.1
py-linq==0.3.0
Pygments==2.0.2
python-dateutil==2.1
python-Levenshtein==0.12.0
python-memcached==1.48
redis==2.10.5
six==1.10.0
slugify==0.0.1
South==0.7.6
sqlparse==0.1.18
titlecase==0.8.1
urllib3==1.14
xlrd==0.9.4
yolk==0.4.3
